requirements.txt
beautifulsoup4
requests
html5lib
discord
python-dotenv
schedule